feat(billing): owner self-serve renewals + grace banner
